Better re-capitalization
Some of the patterns require replacement of a token with multiple letters.
This requires we do something to match the alphabetic case of the source text;
That is "Aa" becomes "Oo" or "AA" into "OO". The previous heuristic was not
terribly effective so now we iterate the characters of the entire token and
actively check the case, setting the output character appropriately. This is
not terribly efficient but it works.